latest code

+++ controllers/category.js
@@ -166,13 +166,15 @@ const deleteCategory = async (req, res) => {
 				return res.status(404).json({ error: 'Category not found' });
 			}
 			else{
-				ImageToDelet = documentExists.image;
-				const imageRes = await deleteFile(ImageToDelet);
-				// console.log("imageRes", imageRes)
+				if(documentExists.image) {
+					ImageToDelet = documentExists.image;
+					const imageRes = await deleteFile(ImageToDelet);
+					// console.log("imageRes", imageRes)
+				}
 			}
 			res.json({ message: 'Category deleted successfully', deletedCategory });
 		} else {
-			console.log('No document found to delete.');
+			res.status(500).json({ error: 'No document found to delete.' });
 		}
 		
 	} catch (err) {
